supoort install hera in arm automatically

+++ scripts/install_artifacts_arm_host.sh
@@ -8,12 +8,45 @@
 
 cd $(dirname "$0")
 
+helpFunction()
+{
+   echo ""
+   echo "use -a means install hera in arm automatically"
+   echo "use -s specify s32v_sdk directory"
+   echo "use -i specify hera installation directory"
+   echo "use -h help information"
+   echo ""
+   exit 1
+}
+
+auto_install=false
+s32v_sdk_dir="/opt/fsl-auto/2.5.2/"
+hera_install_dir="/opt/s32v/hera/"
+
+while getopts "a:s:i:h" opt
+do
+   case "$opt" in
+      a ) auto_install=true ;;
+      s ) s32v_sdk_dir="$OPTARG" ;;
+      i ) hera_install_dir="$OPTARG" ;;
+      h ) helpFunction ;;
+      ? ) helpFunction ;;
+   esac
+done
+
+echo ${auto_install}
+
 if [ ! -d "bin" ] || [ ! -d "lib" ] || [ ! -d "include" ] || [ -d ".git" ]; then
     echo "This script is not designed to be executed directly in this repo"
     exit 1
 fi
 
 echo "This script is for installing cross libhera(arm) to a host linux(amd64) machine"
+
+if [ ${auto_install} == true ]; then
+    unset LD_LIBRARY_PATH
+    source ${s32v_sdk_dir}environment-setup-aarch64-fsl-linux
+fi
 
 if [[ $ARCH == aarch64 ]] || [[ $ARCH == arm* ]]; then
     echo "ARM-Crossplatform toolset activated"
@@ -25,9 +58,11 @@ else
 fi
 
 echo "Input install prefix for cross libhera(arm)"
-install_path="/opt/s32v/hera"
-read -e -i "$install_path" -p "Please input install prefix for libhera: " ans
-install_path="${ans:-$install_path}"
+install_path=${hera_install_dir}
+if [ ${auto_install} == false ]; then
+    read -e -i "$install_path" -p "Please input install prefix for libhera: " ans
+    install_path="${ans:-$install_path}"
+fi
 echo
 
 sudo mkdir -p $install_path/bin
@@ -55,17 +90,27 @@ cat shared/FindHera.cmake >>shared/FindHeraArm.cmake
 echo "CMake module file is shared/FindHeraArm.cmake"
 echo
 
-read -p "Install CMake Package (y/N): " ans
-if [[ $ans = [yY] ]]; then
+if [ ${auto_install} == true ]; then
     echo 'message("${CMAKE_ROOT}/Modules")' >tmp_cmake_root_cmake
     cmake_module_path=$(cmake -N -P tmp_cmake_root_cmake 2>&1)
     rm tmp_cmake_root_cmake
 
-    read -e -i "$cmake_module_path" -p "Please confirm install prefix for CMake module: " ans
-    cmake_module_path="${ans:-$cmake_module_path}"
     echo "Installating CMake Package"
     sudo cp shared/FindHeraArm.cmake $cmake_module_path
+else
+    read -p "Install CMake Package (y/N): " ans
+    if [[ $ans = [yY] ]]; then
+        echo 'message("${CMAKE_ROOT}/Modules")' >tmp_cmake_root_cmake
+        cmake_module_path=$(cmake -N -P tmp_cmake_root_cmake 2>&1)
+        rm tmp_cmake_root_cmake
+
+        read -e -i "$cmake_module_path" -p "Please confirm install prefix for CMake module: " ans
+        cmake_module_path="${ans:-$cmake_module_path}"
+        echo "Installating CMake Package"
+        sudo cp shared/FindHeraArm.cmake $cmake_module_path
+    fi
 fi
+
 echo
 
 echo "Installation Completed"
